feat: Add frontend for recipe/user search
The search is in a form since searching will redirect to a new page and
this is an easy way to pass info to the new page.
The search is in a table so the drop down for search type, input, and
search button can be on the same line.