feat: Search recipe table
Change recipeSearch function to recipeRedirect to better reflect what
the function is doing.

Store what the user searched for with 2 global
variables: searchType and searchText. These variables are used in the
recipeSearch function to search the recipe table for the recipes.
recipeSearch is called when the user is redirected to recipe search
page. Search for recipes based on the search type and text and
pass these found recipes back to the recipe search page.